+++ NEWS.md
@@ -1,5 +1,7 @@
 # devtools 1.X
 
+* `create()` now checks that the directory name is a valid package name (#610).
+
 * `release(check = TRUE)` now runs some additional custom checks. These include:
   
   * Checking that you don't depend on a development version of a package.